+++ source/features/unload_success.py
@@ -9,7 +9,7 @@ class UnloadSuccess(QWidget):
         super(UnloadSuccess, self).__init__()
         self.filePicked = ''
 
-        loadUi("source/program/driver/features/ui/unloadpage_success.ui", self)
+        loadUi("source/features/ui/unloadpage_success.ui", self)
         self.label.setText(f'Successfully removed file {fileN}')
         self.cancel_button.clicked.connect(self.close_window)
         theme = Theme(getTheme())
